fix(agent): the Windows uninstall test, and two files over the 500-line cap
CI caught what my machine could not. `uninstall_unregisters_the_task_before_
deleting_its_definition` is `#[cfg(target_os = "windows")]`, so it does not
compile on this box and passed here while failing on the Windows runner.

It asserted `calls()[0]` contains `Unregister-ScheduledTask` -- a positional
proxy for the property its name states, which is about ordering against DELETING
the definition, not about being first overall. Adding the stop step made the
proxy false while the stated property stayed true. The test now pins what was
meant: stop precedes unregister, unregister precedes removing the XML. That is
strictly more of the contract than the index check covered, and it fails if
either order is broken.

Two files also crossed the 500-line cap, both split along seams the tree already
uses:

- `daemon.rs` (533) -> `daemon/join.rs`. Not an arbitrary cut: serving an
  UNPINNED caller is a separate conversation from serving a peer -- it may pair
  and nothing else -- so both of its bounds and its attempt charge now live
  together, next to the comment explaining what the charge gets wrong.
- `service/tests.rs` (515) -> `service/tests/systemd.rs`, beside the existing
  `windows.rs`, carrying the two tests about the systemd activate/verify
  contract -- including the one this branch added.

Workspace: 641 tests pass, clippy clean, fmt clean, and no file over 500 lines
when checked the way CI checks it.
